move a bunch of stuff around
